complete refactor code
