+++ .gitignore
@@ -6,3 +6,4 @@ package_lock.json
 keys.ts
 .vscode
 /utils/secretKeys.ts
+logs
