+++ server/index.ts
@@ -1,34 +1,26 @@
 import { ConfigureApollo } from '../apollo';
 import express, { Request, Response, NextFunction } from 'express';
 import { ConfigureREST } from '../rest';
-import '../rest/middleware/auth';
+import '../src/api/middlewares/passport-auth';
 import { ConfigureLogger } from '../services/logger';
-import { sendNotification, sendText } from '../services/notification';
 import { startNotificationHandler } from '../services/notificationHandler';
-import { authentication } from '../rest/routes/authentication';
-import { ResetPassword } from '../rest/routes/resetPassword';
-import { refreshToken } from '../rest/routes/refreshToken';
-import { emailsVerification } from '../rest/routes/verifyEmail';
-import { PhoneVerification } from '../rest/routes/verifyPhone';
+import { authentication } from '../src/api/routes/authRoute';
+import { refreshToken } from '../src/api/routes/tokenRefreshRoute';
+import { userdata } from '../src/api/routes/userDataRoute';
+import { verification } from '../src/api/routes/verificationRoute';
 
 const app = express();
-
-// sendNotification('farrukh.faizy@corona-school.de', {
-//     Subject: 'Test Message',
-//     Message: 'Hello from Mailjet # 1',
-// });
 
 app.use(express.json());
 ConfigureLogger();
 ConfigureApollo(app);
 ConfigureREST(app);
 authentication(app);
-ResetPassword(app);
 refreshToken(app);
-emailsVerification(app);
-PhoneVerification(app);
+userdata(app);
+verification(app);
 
-startNotificationHandler(10000);
+startNotificationHandler(30000);
 
 app.use((req, res, next) => {
     const err = new Error('Not found');
